requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
WTForms==3.1.1
Werkzeug==3.0.1
PyPDF2==3.0.1
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0